requirements.txt
# RSS Feed ETL dependencies
feedparser>=6.0.10
html2text>=2020.1.16
pandas>=2.0.0
gspread>=5.0.0
oauth2client>=4.1.2
beautifulsoup4>=4.9.3
pytz>=2021.1
python-dotenv>=1.0.0
PyYAML>=6.0.0

# Resume Matcher dependencies (for future implementation)
PyPDF2>=3.0.0
python-docx>=1.1.0
requests>=2.31.0
scikit-learn>=1.0.0
spacy